+++ test/abstract_server.js
@@ -65,25 +65,6 @@ module.exports = function(moscaSettings, createConnection) {
 	    });
 	  }
 
-	  function buildOpts() {
-	    return {
-	      keepalive: 1000,
-	      clientId: 'mosca_' + require("crypto").randomBytes(8).toString('hex'),
-	      protocolId: 'MQIsdp',
-	      protocolVersion: 3
-	    };
-	  }
-
-
-	  function donner(count, done) {
-	    return function() {
-	      count--;
-	      if (count === 0) {
-	        done();
-	      }
-	    };
-	  }
-
 	  it("should pass itself in the callback", function(done) {
 	    secondInstance = new mosca.Server(moscaSettings(), function(err, server) {
 	      expect(server === secondInstance).to.be.true;
@@ -539,37 +520,6 @@ module.exports = function(moscaSettings, createConnection) {
 	    ]);
 	  });
 
-	  it("should pass the backend settings to ascoltatori.build", function(done) {
-	    var spy = sinon.spy(ascoltatori, "build");
-	    var newSettings = moscaSettings();
-
-	    newSettings.backend = {
-	      type: "mqtt",
-	      json: false,
-	      port: settings.port,
-	      keepalive: 3000,
-	      host: "127.0.0.1",
-	      mqtt: require("mqtt")
-	    };
-
-	    var server = new mosca.Server(newSettings);
-
-	    async.series([
-
-	      function(cb) {
-	        server.on("ready", cb);
-	      },
-
-	      function(cb) {
-	        expect(spy).to.have.been.calledWith(newSettings.backend);
-	        cb();
-	      },
-
-	      function(cb) {
-	        server.close(cb);
-	      }
-	    ], done);
-	  });
 
 	  it("should support subscribing to wildcards", function(done) {
 	    var d = donner(2, done);
@@ -635,173 +585,6 @@ module.exports = function(moscaSettings, createConnection) {
 	    });
 	  });
 
-	  it("should support subscribing correctly to wildcards in a tree-based topology", function(done) {
-	    var d = donner(3, done);
-
-	    async.waterfall([
-
-	      function(cb) {
-	        settings.backend = {
-	          port: settings.port,
-	          type: "mqtt"
-	        };
-	        settings.port = nextPort();
-	        secondInstance = new mosca.Server(settings, function() {
-	          cb();
-	        });
-	      },
-
-	      function(cb) {
-	        buildAndConnect(d, function(client1) {
-	          cb(null, client1);
-	        });
-	      },
-
-	      function(client1, cb) {
-	        var called = false;
-	        client1.on("publish", function(packet) {
-	          expect(called).to.be.eql(false);
-	          called = true;
-	          setTimeout(function() {
-	            client1.disconnect();
-	          });
-	        });
-
-	        var subscriptions = [{
-	            topic: "hello/#",
-	            qos: 0
-	          }
-	        ];
-	        client1.subscribe({
-	          subscriptions: subscriptions,
-	          messageId: 42
-	        });
-
-	        client1.on("suback", function() {
-	          cb(null);
-	        });
-	      },
-
-	      function(cb) {
-	        buildAndConnect(d, function(client3) {
-	          cb(null, client3);
-	        });
-	      },
-
-	      function(client3, cb) {
-	        var subscriptions = [{
-	            topic: "hello/#",
-	            qos: 0
-	          }
-	        ];
-	        client3.subscribe({
-	          subscriptions: subscriptions,
-	          messageId: 42
-	        });
-	        client3.on("suback", function() {
-	          // we need to simulate a "stuck" subscription
-	          client3.stream.end();
-	          cb(null);
-	        });
-	      },
-
-	      function(cb) {
-	        buildAndConnect(d, function(client2) {
-	          cb(null, client2);
-	        });
-	      },
-
-	      function(client2, cb) {
-	        client2.publish({
-	          topic: "hello/world",
-	          payload: "some data"
-	        });
-	        client2.disconnect();
-	      }
-	    ]);
-	  });
-
-	  it("should not wrap messages with \"\" in a tree-based topology", function(done) {
-	    var d = donner(2, done);
-
-	    async.waterfall([
-
-	      function(cb) {
-	        buildAndConnect(d, function(client1) {
-	          cb(null, client1);
-	        });
-	      },
-
-	      function(client1, cb) {
-	        client1.on("publish", function(packet) {
-	          expect(packet.payload).to.be.eql("some data");
-	          client1.disconnect();
-	        });
-
-	        var subscriptions = [{
-	            topic: "hello/#",
-	            qos: 0
-	          }
-	        ];
-
-	        client1.subscribe({
-	          subscriptions: subscriptions,
-	          messageId: 42
-	        });
-	        client1.on("suback", function() {
-	          cb(null);
-	        });
-	      },
-
-	      function(cb) {
-	        settings.backend = {
-	          port: settings.port,
-	          type: "mqtt"
-	        };
-	        settings.port = settings.port + 1000;
-	        secondInstance = new mosca.Server(settings, function() {
-	          cb();
-	        });
-	      },
-
-	      function(cb) {
-	        buildAndConnect(d, function(client2) {
-	          cb(null, client2);
-	        });
-	      },
-
-	      function(client2, cb) {
-	        client2.publish({
-	          topic: "hello/world",
-	          payload: "some data"
-	        });
-	        client2.disconnect();
-	      }
-	    ]);
-	  });
-	  
-	  it("should support specifying an Ascoltatore instead of backend options", function(done) {
-
-	    async.waterfall([
-
-	      function(cb) {
-	        settings.ascoltatore = ascoltatori.build({
-	          json: false
-	        });
-	        settings.port = settings.port + 1000;
-	        mosca.Server(settings, cb);
-	      },
-
-	      function(server, cb) {
-	        secondInstance = server;
-	        buildAndConnect(done, function(client) {
-	          client.disconnect();
-	          cb(null);
-	        });
-	      }
-
-	    ]);
-	  });
 
 	  it("should support unsubscribing a single client", function(done) {
 	    var d = donner(2, done);
